requirements.txt
# Requirements for Stock Market AI Agent
streamlit
yfinance
pandas-ta
plotly
pandas
numpy
requests
pymongo
python-dotenv
pytz
psycopg2-binary
sqlalchemy
